+++ HISTORY.rst
@@ -6,6 +6,8 @@ Changelog
 
 * Fix formatting of form field error output when there are multiple field errors
 
+* Fix bug in form validation signaling required field missing on choices where the value is falsy, e.g. 0
+
 
 4.0.0 (2016-09-14)
 ~~~~~~~~~~~~~~~~~~